prettier issue fixed.

+++ contracts/oracles/default/UniswapV3PriceOracle.sol
@@ -139,6 +139,6 @@ contract UniswapV3PriceOracle is IPriceOracle {
   }
 
   function getPriceX96FromSqrtPriceX96(uint160 sqrtPriceX96) public pure returns (uint256 priceX96) {
-    return FullMath.mulDiv(sqrtPriceX96, sqrtPriceX96, uint(2**(96*2)) / 1e18);
+    return FullMath.mulDiv(sqrtPriceX96, sqrtPriceX96, uint256(2**(96 * 2)) / 1e18);
   }
 }
